fix(pvk): exclude iverksatte tiltak from "uten frist" filter and align PVO counts
- Exclude tiltak marked as iverksatt from the "uten frist" list/count
- Re-apply active tiltak filter after editing a tiltak so the list updates
  immediately (e.g. when marking as iverksatt) without a page refresh
- Align PVO view's "uten ansvarlig"/"uten frist" counts and filters with
  the read-only view so tab counts match the displayed list
- Extract shared tiltakFilterValues/filterTiltakList helpers to
  tiltakConstants.ts to avoid duplicated filter logic
