+++ apps/frontend-nextjs/app/constants/etterlevelseDokumentasjon/personvernkonsekvensevurdering/tiltak/tiltakConstants.ts
@@ -14,3 +14,23 @@ export interface ITiltak extends IDomainObject {
   iverksattDato?: string
   iverksettingsKommentar?: string
 }
+
+export const tiltakFilterValues = {
+  alleTiltak: 'alleTiltak',
+  utenAnsvarlig: 'utenAnsvarlig',
+  utenFrist: 'utenFrist',
+}
+
+export const filterTiltakList = (tiltakList: ITiltak[], filter: string): ITiltak[] => {
+  switch (filter) {
+    case tiltakFilterValues.utenAnsvarlig:
+      return tiltakList.filter(
+        (tiltak: ITiltak) =>
+          !tiltak.ansvarlig || (!tiltak.ansvarlig.navIdent && !tiltak.ansvarligTeam.name)
+      )
+    case tiltakFilterValues.utenFrist:
+      return tiltakList.filter((tiltak: ITiltak) => !tiltak.iverksatt && !tiltak.frist)
+    default:
+      return tiltakList
+  }
+}
